(wip) refactor e2e-helper, split this huge class to multiple classes
