+++ src/e2e-helper/e2e-command-helper.js
@@ -1,0 +1,308 @@
+// @flow
+import rightpad from 'pad-right';
+import chalk from 'chalk';
+import path from 'path';
+import childProcess from 'child_process';
+import { expect } from 'chai';
+import { NOTHING_TO_TAG_MSG } from '../cli/commands/public-cmds/tag-cmd';
+import { removeChalkCharacters } from '../utils';
+import runInteractive from '../interactive/utils/run-interactive-cmd';
+import type { InteractiveInputs } from '../interactive/utils/run-interactive-cmd';
+import ScopesData from './e2e-scopes';
+
+const DEFAULT_DEFAULT_INTERVAL_BETWEEN_INPUTS = 200;
+
+export default class CommandHelper {
+  scopes: ScopesData;
+  debugMode: boolean;
+  bitBin: string;
+  constructor(scopes: ScopesData, debugMode: boolean) {
+    this.scopes = scopes;
+    this.debugMode = debugMode;
+    this.bitBin = process.env.npm_config_bit_bin || 'bit'; // e.g. npm run e2e-test --bit_bin=bit-dev
+  }
+
+  runCmd(cmd: string, cwd: string = this.scopes.localScopePath): string {
+    if (this.debugMode) console.log(rightpad(chalk.green('cwd: '), 20, ' '), cwd); // eslint-disable-line no-console
+    if (cmd.startsWith('bit ')) cmd = cmd.replace('bit', this.bitBin);
+    if (this.debugMode) console.log(rightpad(chalk.green('command: '), 20, ' '), cmd); // eslint-disable-line no-console
+    // $FlowFixMe
+    const cmdOutput = childProcess.execSync(cmd, { cwd, shell: true });
+    if (this.debugMode) console.log(rightpad(chalk.green('output: '), 20, ' '), chalk.cyan(cmdOutput.toString())); // eslint-disable-line no-console
+    return cmdOutput.toString();
+  }
+
+  listRemoteScope(raw: boolean = true, options: string = '') {
+    return this.runCmd(`bit list ${this.scopes.remoteScope} ${options} ${raw ? '--raw' : ''}`);
+  }
+  listLocalScope(options: string = '') {
+    return this.runCmd(`bit list ${options}`);
+  }
+  listLocalScopeParsed(options: string = '') {
+    const output = this.runCmd(`bit list --json ${options}`);
+    return JSON.parse(output);
+  }
+  listRemoteScopeParsed(options: string = '') {
+    const output = this.runCmd(`bit list ${this.scopes.remoteScope} --json ${options}`);
+    return JSON.parse(output);
+  }
+
+  catScope(includeExtraData: boolean = false) {
+    const extraData = includeExtraData ? '--json-extra' : '';
+    const result = this.runCmd(`bit cat-scope --json ${extraData}`);
+    return JSON.parse(result);
+  }
+
+  catObject(hash: string, parse: boolean = false) {
+    const result = this.runCmd(`bit cat-object ${hash}`);
+    if (!parse) return result;
+    return JSON.parse(result);
+  }
+
+  catComponent(id: string, cwd?: string): Object {
+    const result = this.runCmd(`bit cat-component ${id}`, cwd);
+    return JSON.parse(result);
+  }
+  addComponent(filePaths: string, options: Object = {}, cwd: string = this.scopes.localScopePath) {
+    const value = Object.keys(options)
+      .map(key => `-${key} ${options[key]}`)
+      .join(' ');
+    return this.runCmd(`bit add ${filePaths} ${value}`, cwd);
+  }
+  getConfig(configName: string) {
+    return this.runCmd(`bit config get ${configName}`);
+  }
+  delConfig(configName: string) {
+    return this.runCmd(`bit config del ${configName}`);
+  }
+  setConfig(configName: string, configVal: string) {
+    return this.runCmd(`bit config set ${configName} ${configVal}`);
+  }
+  untrackComponent(id: string = '', all: boolean = false, cwd: string = this.scopes.localScopePath) {
+    return this.runCmd(`bit untrack ${id} ${all ? '--all' : ''}`, cwd);
+  }
+  removeComponent(id: string, flags: string = '') {
+    return this.runCmd(`bit remove ${id} ${flags}`);
+  }
+  deprecateComponent(id: string, flags: string = '') {
+    return this.runCmd(`bit deprecate ${id} ${flags}`);
+  }
+  undeprecateComponent(id: string, flags: string = '') {
+    return this.runCmd(`bit undeprecate ${id} ${flags}`);
+  }
+  tagComponent(id: string, tagMsg: string = 'tag-message', options: string = '') {
+    return this.runCmd(`bit tag ${id} -m ${tagMsg} ${options}`);
+  }
+  tagWithoutMessage(id: string, version: string = '', options: string = '') {
+    return this.runCmd(`bit tag ${id} ${version} ${options}`);
+  }
+  tagAllComponents(options: string = '', version: string = '', assertTagged: boolean = true) {
+    const result = this.runCmd(`bit tag -a ${version} ${options} `);
+    if (assertTagged) expect(result).to.not.have.string(NOTHING_TO_TAG_MSG);
+    return result;
+  }
+  tagScope(version: string, message: string = 'tag-message', options: string = '') {
+    return this.runCmd(`bit tag -s ${version} -m ${message} ${options}`);
+  }
+
+  untag(id: string) {
+    return this.runCmd(`bit untag ${id}`);
+  }
+
+  exportComponent(id: string, scope: string = this.scopes.remoteScope, assert: boolean = true) {
+    const result = this.runCmd(`bit export ${scope} ${id}`);
+    if (assert) expect(result).to.not.have.string('nothing to export');
+    return result;
+  }
+
+  ejectComponents(ids: string, flags?: string) {
+    return this.runCmd(`bit eject ${ids} ${flags || ''}`);
+  }
+  ejectComponentsParsed(ids: string, flags?: string) {
+    const result = this.runCmd(`bit eject ${ids} ${flags || ''} --json`);
+    const jsonStart = result.indexOf('{');
+    const jsonResult = result.substring(jsonStart);
+    return JSON.parse(jsonResult);
+  }
+
+  exportAllComponents(scope: string = this.scopes.remoteScope) {
+    return this.runCmd(`bit export ${scope}`);
+  }
+
+  importComponent(id: string) {
+    return this.runCmd(`bit import ${this.scopes.remoteScope}/${id}`);
+  }
+
+  importManyComponents(ids: string[]) {
+    const idsWithRemote = ids.map(id => `${this.scopes.remoteScope}/${id}`);
+    return this.runCmd(`bit import ${idsWithRemote.join(' ')}`);
+  }
+
+  importComponentWithOptions(id: string = 'bar/foo.js', options: Object) {
+    const value = Object.keys(options)
+      .map(key => `-${key} ${options[key]}`)
+      .join(' ');
+    return this.runCmd(`bit import ${this.scopes.remoteScope}/${id} ${value}`);
+  }
+
+  importAllComponents(writeToFileSystem: boolean = false) {
+    return this.runCmd(`bit import ${writeToFileSystem ? '--merge' : ''}`);
+  }
+
+  isolateComponent(id: string, flags: string): string {
+    const isolatedEnvOutput = this.runCmd(
+      `bit isolate ${this.scopes.remoteScope}/${id} ${this.scopes.remoteScopePath} ${flags}`
+    );
+    const isolatedEnvOutputArray = isolatedEnvOutput.split('\n').filter(str => str);
+    return isolatedEnvOutputArray[isolatedEnvOutputArray.length - 1];
+  }
+
+  importExtension(id: string) {
+    return this.runCmd(`bit import ${id} --extension`);
+  }
+
+  build(id?: string = '') {
+    return this.runCmd(`bit build ${id}`);
+  }
+
+  buildComponentWithOptions(id: string = '', options: Object, cwd: string = this.scopes.localScopePath) {
+    const value = Object.keys(options)
+      .map(key => `-${key} ${options[key]}`)
+      .join(' ');
+    return this.runCmd(`bit build ${id} ${value}`, cwd);
+  }
+
+  testComponent(id: string = '') {
+    return this.runCmd(`bit test ${id}`);
+  }
+
+  testComponentWithOptions(id: string = '', options: Object, cwd: string = this.scopes.localScopePath) {
+    const value = Object.keys(options)
+      .map(key => `-${key} ${options[key]}`)
+      .join(' ');
+    return this.runCmd(`bit test ${id} ${value}`, cwd);
+  }
+
+  status() {
+    return this.runCmd('bit status');
+  }
+
+  statusJson() {
+    const status = this.runCmd('bit status --json');
+    return JSON.parse(status);
+  }
+
+  showComponent(id: string = 'bar/foo') {
+    return this.runCmd(`bit show ${id}`);
+  }
+
+  showComponentParsed(id: string = 'bar/foo') {
+    const output = this.runCmd(`bit show ${id} --json`);
+    return JSON.parse(output);
+  }
+
+  showComponentWithOptions(id: string = 'bar/foo', options: Object) {
+    const value = Object.keys(options)
+      .map(key => `-${key} ${options[key]}`)
+      .join(' ');
+    return this.runCmd(`bit show ${id} ${value}`);
+  }
+
+  checkoutVersion(version: string, ids: string, flags?: string, cwd?: string) {
+    return this.runCmd(`bit checkout ${version} ${ids} ${flags || ''}`, cwd);
+  }
+
+  checkout(values: string) {
+    return this.runCmd(`bit checkout ${values}`);
+  }
+
+  mergeVersion(version: string, ids: string, flags?: string) {
+    return this.runCmd(`bit merge ${version} ${ids} ${flags || ''}`);
+  }
+
+  diff(id?: string = '') {
+    const output = this.runCmd(`bit diff ${id}`);
+    return removeChalkCharacters(output);
+  }
+
+  move(from: string, to: string) {
+    return this.runCmd(`bit move ${path.normalize(from)} ${path.normalize(to)}`);
+  }
+  ejectConf(id: string = 'bar/foo', options: ?Object) {
+    const value = options
+      ? Object.keys(options) // $FlowFixMe
+        .map(key => `-${key} ${options[key]}`)
+        .join(' ')
+      : '';
+    return this.runCmd(`bit eject-conf ${id} ${value}`);
+  }
+  injectConf(id: string = 'bar/foo', options: ?Object) {
+    const value = options
+      ? Object.keys(options) // $FlowFixMe
+        .map(key => `-${key} ${options[key]}`)
+        .join(' ')
+      : '';
+    return this.runCmd(`bit inject-conf ${id} ${value}`);
+  }
+  doctor(options: Object) {
+    const parsedOpts = this.parseOptions(options);
+    return this.runCmd(`bit doctor ${parsedOpts}`);
+  }
+
+  doctorOne(diagnosisName: string, options: Object, cwd?: string) {
+    const parsedOpts = this.parseOptions(options);
+    return this.runCmd(`bit doctor "${diagnosisName}" ${parsedOpts}`, cwd);
+  }
+
+  doctorList(options: Object) {
+    const parsedOpts = this.parseOptions(options);
+    return this.runCmd(`bit doctor --list ${parsedOpts}`);
+  }
+
+  doctorJsonParsed() {
+    const result = this.runCmd('bit doctor --json');
+    return JSON.parse(result);
+  }
+
+  parseOptions(options: Object): string {
+    const value = Object.keys(options)
+      .map((key) => {
+        const keyStr = key.length === 1 ? `-${key}` : `--${key}`;
+        return `${keyStr} ${options[key]}`;
+      })
+      .join(' ');
+    return value;
+  }
+
+  async runInteractiveCmd({
+    args = [],
+    inputs = [],
+    // Options for the process (execa)
+    processOpts = {
+      cwd: this.scopes.localScopePath
+    },
+    // opts for interactive
+    opts = {
+      defaultIntervalBetweenInputs: DEFAULT_DEFAULT_INTERVAL_BETWEEN_INPUTS,
+      verbose: false
+    }
+  }: {
+    args: string[],
+    inputs: InteractiveInputs,
+    processOpts: Object,
+    opts: {
+      // Default interval between inputs in case there is no specific interval
+      defaultIntervalBetweenInputs: number,
+      verbose: boolean
+    }
+  }) {
+    const processName = this.bitBin || 'bit';
+    opts.verbose = !!this.debugMode;
+    const { stdout } = await runInteractive({ processName, args, inputs, processOpts, opts });
+    if (this.debugMode) {
+      console.log(rightpad(chalk.green('output: \n'), 20, ' ')); // eslint-disable-line no-console
+      console.log(chalk.cyan(stdout)); // eslint-disable-line no-console
+    }
+    return stdout;
+  }
+}
